moved LivePPManager out of mason.lib and removed the headers
